Connect BookList to the store and update presentational, container components
