+++ src/components/App.js
@@ -1,31 +1,12 @@
 import React from 'react';
-import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
 import BookList from '../containers/BooksList';
 import BooksForm from '../containers/BooksForm';
-import { createBook, deleteBook } from '../actions';
 
-const App = ({ books }) => {
+const App = () => (
   <div className="App">
-    <BookList books={books} />
+    <BookList />
     <BooksForm />
-  </div>;
-};
+  </div>
+);
 
-App.propTypes = {
-  books: PropTypes.arrayOf(
-    PropTypes.shape({
-      id: PropTypes.number.isRequired,
-      title: PropTypes.string.isRequired,
-      category: PropTypes.string.isRequired,
-    }),
-  ).isRequired,
-};
-
-const mapStateToProps = (state) => ({ books: state.books });
-const mapDispatchToProps = (dispatch) => ({
-  create: (book) => dispatch(createBook(book)),
-  delete: (book) => dispatch(deleteBook(book)),
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default App;
